переписал компоненты и страницы на typeScript
