--- src/app/auth/register/page.jsx
@@ -1,131 +1,0 @@
-'use client';
-
-import axios from 'axios';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { useForm } from 'react-hook-form';
-
-const RegisterPage = () => {
-	const {
-		register,
-		handleSubmit,
-		formState: { errors },
-	} = useForm();
-	const router = useRouter();
-
-	const onSubmit = handleSubmit(async (data) => {
-		if (data.password !== data.confirmPassword) {
-			return alert('Пароли не совпадают');
-		}
-		const res = await axios.post('http://localhost:3001/user', data);
-
-		if (res.status === 201) {
-			router.push('/');
-		}
-	});
-
-	console.log(errors);
-
-	return (
-		<div className='h-[calc(100vh-7rem)] flex justify-center items-center'>
-			<form
-				onSubmit={onSubmit}
-				className='w-1/4  border-2 border-solid border-white p-5'
-			>
-				<h2 className='text-slate-200 font-bold text-2xl mb-4 text-center'>
-					Регистрация
-				</h2>
-
-				<label htmlFor='username' className='text-slate-500 mb-2 block text-sm'>
-					Username:
-				</label>
-				<input
-					type='text'
-					{...register('username', {
-						required: {
-							value: true,
-							message: 'Имя пользователя не введено',
-						},
-					})}
-					className='p-3 rounded block mb-2 bg-slate-900 text-slate-300 w-full'
-					placeholder='Имя пользователя'
-				/>
-				{errors.username && (
-					<span className='text-red-500 text-xs'>
-						{errors.username.message}
-					</span>
-				)}
-
-				<label htmlFor='email' className='text-slate-500 mb-2 block text-sm'>
-					Email:
-				</label>
-				<input
-					type='email'
-					{...register('email', {
-						required: {
-							value: true,
-							message: 'Электронная почта не введена',
-						},
-					})}
-					className='p-3 rounded block mb-2 bg-slate-900 text-slate-300 w-full'
-					placeholder='Ваш e-mail'
-				/>
-				{errors.email && (
-					<span className='text-red-500 text-xs'>{errors.email.message}</span>
-				)}
-
-				<label htmlFor='password' className='text-slate-500 mb-2 block text-sm'>
-					Password:
-				</label>
-				<input
-					type='password'
-					{...register('password', {
-						required: {
-							value: true,
-							message: 'Пароль не введен',
-						},
-					})}
-					className='p-3 rounded block mb-2 bg-slate-900 text-slate-300 w-full'
-					placeholder='********'
-				/>
-				{errors.password && (
-					<span className='text-red-500 text-sm'>
-						{errors.password.message}
-					</span>
-				)}
-
-				<label
-					htmlFor='confirmPassword'
-					className='text-slate-500 mb-2 block text-sm'
-				>
-					Confirm Password:
-				</label>
-				<input
-					type='password'
-					{...register('confirmPassword', {
-						required: {
-							value: true,
-							message: 'Пароль не введен',
-						},
-					})}
-					className='p-3 rounded block mb-2 bg-slate-900 text-slate-300 w-full'
-					placeholder='********'
-				/>
-				{errors.confirmPassword && (
-					<span className='text-red-500 text-sm'>
-						{errors.confirmPassword.message}
-					</span>
-				)}
-
-				<button className='w-full bg-blue-500 text-white p-3 rounded-lg mt-2'>
-					Регистрация
-				</button>
-				<span className='cursor-pointer text-white block mt-4'>
-					У вас уже есть аккаунт? <Link href='/auth/login'> ВОЙТИ </Link>
-				</span>
-			</form>
-		</div>
-	);
-};
-
-export default RegisterPage;
