+++ src/components/MenuItem/MenuItem.tsx
@@ -1,0 +1,33 @@
+'use client';
+import React from 'react';
+import './menuItem.css';
+
+interface MenuItemProps {
+  title: string;
+  price: string;
+  tags: string;
+}
+
+const MenuItem: React.FC<MenuItemProps> = ({ title, price, tags }) => (
+  <div className='app__menuitem'>
+    <div className='app__menuitem-head'>
+      <div className='app__menuitem-name'>
+        <p className='p__cormorant' style={{ color: '#DCCA87' }}>
+          {title}
+        </p>
+      </div>
+      <div className='app__menuitem-dash' />
+      <div className='app__menuitem-price'>
+        <p className='p__cormorant'>{price}</p>
+      </div>
+    </div>
+
+    <div className='app__menuitem-sub'>
+      <p className='p__opensans' style={{ color: '#AAAAAA' }}>
+        {tags}
+      </p>
+    </div>
+  </div>
+);
+
+export default MenuItem;
